requirements.txt
streamlit==1.36.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
python-dotenv==1.0.1
langchain==0.2.7
langchain-core==0.2.18
langchain-community==0.2.7
langchain-google-genai==1.0.7
chromadb==0.5.4
youtube_transcript_api==0.6.2
pathlib==1.0.1
faiss-cpu==1.8.0.post1